run collection do work submission

+++ src/Applications/use_case/DeleteCommentUseCase.js
@@ -5,11 +5,13 @@ class DeleteCommentUseCase {
   }
 
   async execute(useCasePayload) {
-    const {commentId, threadId} = useCasePayload;
+    const {commentId, threadId, owner} = useCasePayload;
     await this._commentRepository.verifyThreadId(threadId);
     await this._commentRepository.verifyCommentId(commentId);
+    await this._commentRepository.verifyOwner(owner, commentId);
     return this._commentRepository.deleteComment(commentId);
   }
+
   _validatePayload(payload) {}
 }
